# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления арендой автомобилей (CRM).",
        "title": "Auto Hub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Возвращает JWT токен по email и паролю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Войти в систему",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cars": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Список машин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (FREE, RENTED, REPAIR, SOLD)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по номеру (ILIKE)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CarsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Создать машину",
                "parameters": [
                    {
                        "description": "Данные машины",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Car"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Детали машины",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID машины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CarDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Обновить машину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID машины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные машины",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Car"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Удалить машину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID машины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Car": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.CarStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "domain.CarStatus": {
            "type": "string",
            "enum": [
                "FREE",
                "RENTED",
                "REPAIR",
                "SOLD"
            ],
            "x-enum-varnames": [
                "CarStatusFree",
                "CarStatusRented",
                "CarStatusRepair",
                "CarStatusSold"
            ]
        },
...
        "domain.ContractStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "ContractStatusActive",
                "ContractStatusCompleted",
                "ContractStatusCancelled"
            ]
        },
        "domain.Driver": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.DriverStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.DriverStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "DriverStatusActive",
                "DriverStatusInactive"
            ]
        },
        "domain.Fine": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "car_id": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "fine_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.FineStatus"
                }
            }
        },
        "domain.FineStatus": {
            "type": "string",
            "enum": [
                "UNPAID",
                "PAID",
                "DISPUTED"
            ],
            "x-enum-varnames": [
                "FineStatusUnpaid",
                "FineStatusPaid",
                "FineStatusDisputed"
            ]
        },
        "domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                }
            }
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "PAID",
                "UNPAID",
                "OVERDUE"
            ],
            "x-enum-varnames": [
                "PaymentStatusPaid",
                "PaymentStatusUnpaid",
                "PaymentStatusOverdue"
            ]
        },
        "dto.ActiveContractBrief": {
            "type": "object",
            "properties": {
                "driver": {
                    "$ref": "#/definitions/dto.DriverName"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.CarDetailResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContractForCarDetail"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "fines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Fine"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.CarStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CarWithContracts": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ActiveContractBrief"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.CarStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CarsListResponse": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CarWithContracts"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ContractForCarDetail": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/domain.Driver"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Payment"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ContractStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateCarRequest": {
            "type": "object",
            "required": [
                "brand",
                "model",
                "plate_number",
                "status",
                "vin",
                "year"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer",
                    "minimum": 0
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "FREE",
                        "RENTED",
                        "REPAIR",
                        "SOLD"
                    ]
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 1900
                }
            }
        },
...
        "dto.DriverName": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateCarRequest": {
            "type": "object",
            "required": [
                "brand",
                "model",
                "plate_number",
                "status",
                "vin",
                "year"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer",
                    "minimum": 0
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "FREE",
                        "RENTED",
                        "REPAIR",
                        "SOLD"
                    ]
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 1900
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}